image: allow image to pivot to a second disk
Some aws ec2 instances have local disks attached to them which are meant for
temporary storage or large workloads (there is no option to attach root to these
disks). Since aws ec2 only allows the rootfs to be on the EBS network-attached
storage (or instance-backed disks, but that's not really relevent here).

So include the ospack in the image and check on boot if a second disk is
attached. If so, extract the compressed OS structure there and use that as the
rootfs.

Signed-off-by: Christopher Obbard <chris.obbard@collabora.com>