flasher: be more specific with the path to the block device
The logic for the partition before was quite primitive; we just appended
the partition to the path to the drive. So we always ended up with
${DRIVE}2. Some drives end up with a path like ${DRIVE}p2 or some other
suffix. So don't try to be clever and instead use the output of lsblk to
find the partition path.

Also fix some issue with the partition table not resized to full and
also probing the partition after resizing.

Signed-off-by: Christopher Obbard <chris.obbard@collabora.com>