pivot: rescan partition table & make tar less noisy
A similar racy issue happened with the EFI flasher: after partitioning
the kernel wasn't aware of the new disks. So attempt to force a re-scan.

Also when this does fail the AWS console ends up overflowing with spew
from tar not being able to write to a full root. So quieten up tar's
stderr.

Signed-off-by: Christopher Obbard <chris.obbard@collabora.com>