Limit used memory to 16G
One source of variation in the test setup is the amount
of memory available which affects the swap and testfile
sizes. This can lead to different behaviors depending on
ssds (or just outright failure on machines with huge
amounts of memory).

Since the test is normalizing to 16G of memory, boot
the images with mem=16G kernel parameter such that the
maximum available memory used by the system is 16G.

Signed-off-by: Christopher Obbard <chris.obbard@collabora.com>