qemu-flasher: Add a disk with 4096-byte sectors
As per https://github.com/iocost-benchmark/resctl-demo-image-recipe/issues/35#issuecomment-2238736902
we think that the root cause for the installer not working on certain hardware
is due to modern SSDs having different sector sizes. Add a disk with fixed
sector size to the qemu script to be able to test this case.

Signed-off-by: Christopher Obbard <chris.obbard@collabora.com>